vim: fix ale prettier command
Make sure that the config will trump the CLI when available & merge
options when it's passed from the CLI & it's not in the file.

In my case I want to respect vim's `textwidth` when I format, but if
a project uses a different value in a config file that should be
respected.