vim: back to Files from Gfiles
The issues with `Gfiles` is that it uses `git ls-files` which will only
show files tracked by git. Which means, added new files doesn't show by
default & deleted files are still showing up. Because the first is not
in `git` history & the later is in the `git` history.

I tried to tweak `Gfiles` to use `git ls-files --cached --modifed --deleted --exclude=node_modules`

But now I was seeing duplicate files, because it can be delete & cached
for example still. So I'm back to `Files` which doesn't have any of
these issues & it has the pro of being able to open any file on disk.
Which is useful in the situation of files in the repo that acts as
`config.yml.template` & then locally ignores `config.yml`.

I want to edit `config.yml` but `Gfiles` will only show `config.yml.template`