vim: vim-commentary -> tcomment
I like vim-commentary & its simplicity but the fact that it doesn't support motions is a bummer. tcomment does support that, for example I can do `gcW`

```
function foo(a, |b, c) {}

// it will become

function foo(a, /* b, */ c) {}
```

Uncommenting is a bit buggy, but with vim-surround I can do `ds/` & it will remove the comment. It's not perfect & it's a massive plugin compared to vim-commentary but it's very convenient.