install: update script
This is a result of running the script on a new MBP M2 machine, I think
when I started using nix I didn't understand much. Plus, I guess I had
a "TODO to fix the script later" which of course never happened. So this
was a good moment to fix it. I still have a couple of things to iron up
but they are not urgent yet.